pyproject.toml
[tool.poetry]
name = "epic_events"
version = "0.1.0"
description = ""
authors = ["geo1310 <gbriche59@yahoo.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
SQLAlchemy = "^2.0.30"
psycopg2-binary = "^2.9.9"
passlib = "^1.7.4"
pymysql = "^1.1.0"
cryptography = "^42.0.7"
python-dotenv = "^1.0.1"
bcrypt = "3.2.0"
black = "^24.4.2"
pytest = "^8.2.0"
pycln = "^2.4.0"
flake8 = "^7.0.0"
isort = "^5.13.2"
flake8-pyproject = "^1.2.3"
coverage = "^7.5.1"
locust = "^2.27.0"
pytest-cov = "^5.0.0"
pyjwt = "^2.8.0"
rich = "^13.7.1"
sentry-sdk = "^2.3.1"
pg8000 = "^1.31.2"
colorlog = "^6.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.locust]
locustfile = "app/tests/performance_tests/locustfile.py"
host = "http://127.0.0.1:5000"
users = 6
spawn-rate = 1

[tool.pytest]
testpaths = "app/tests/"

[tool.coverage.run]
branch = true
omit = ["app/tests/*", "__*", "config.py", "app/dev", "init*", "views*"]

[tool.flake8]
max-line-length = 119
exclude = ".venv/, docs/"
extend-ignore = [
    # PEP 8 recommends to treat : in slices as a binary operator with the lowest priority, and to leave an equal
    # amount of space on either side, except if a parameter is omitted (e.g. ham[1 + 1 :]).
    # This behaviour may raise E203 whitespace before ':' warnings in style guide enforcement tools like Flake8.
    # Since E203 is not PEP 8 compliant, we tell Flake8 to ignore this warning.
    # https://black.readthedocs.io/en/stable/the_black_code_style/current_style.html#slices    
    "E203"
]

[tool.black]
line-length = 119
